refactor: avoid duplication
