+++ extensions/api/api-core/src/main/java/org/eclipse/dataspaceconnector/api/transformer/DtoTransformerRegistryImpl.java
@@ -18,13 +18,11 @@ package org.eclipse.dataspaceconnector.api.transformer;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.eclipse.dataspaceconnector.spi.transformer.TransformerContext;
+import org.eclipse.dataspaceconnector.spi.transformer.TransformerContextImpl;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
@@ -68,11 +66,11 @@ public class DtoTransformerRegistryImpl implements DtoTransformerRegistry {
     @Override
     public <INPUT, OUTPUT> Result<OUTPUT> transform(@NotNull INPUT object, @NotNull Class<OUTPUT> outputType) {
 
-        var ctx = new DtoTransformerContext(this);
+        var ctx = new TransformerContextImpl(this);
 
         var output = transform(object, outputType, ctx);
 
-        return ctx.hasProblems() ? Result.failure(ctx.problems) : Result.success(output);
+        return ctx.hasProblems() ? Result.failure(ctx.getProblems()) : Result.success(output);
     }
 
     private <INPUT, OUTPUT> OUTPUT transform(INPUT object, Class<OUTPUT> outputType, TransformerContext context) {
@@ -88,33 +86,4 @@ public class DtoTransformerRegistryImpl implements DtoTransformerRegistry {
         return outputType.cast(((DtoTransformer<INPUT, OUTPUT>) t).transform(object, context));
     }
 
-    public static class DtoTransformerContext implements TransformerContext {
-        private final DtoTransformerRegistry registry;
-        private final List<String> problems;
-
-        public DtoTransformerContext(DtoTransformerRegistry registry) {
-            this.registry = registry;
-            problems = new ArrayList<>();
-        }
-
-        @Override
-        public boolean hasProblems() {
-            return !problems.isEmpty();
-        }
-
-        @Override
-        public List<String> getProblems() {
-            return problems;
-        }
-
-        @Override
-        public void reportProblem(String problem) {
-            problems.add(problem);
-        }
-
-        @Override
-        public <INPUT, OUTPUT> @Nullable OUTPUT transform(INPUT object, Class<OUTPUT> outputType) {
-            return registry.transform(object, outputType).getContent();
-        }
-    }
 }
